+++ CHANGELOG.md
@@ -1,3 +1,16 @@
+## 0.0.13
+
+- fix: improve path sanitization
+- feat: support custom log callback
+
+## 0.0.12
+
+- fix: expose pathAliases on RouteId constructor
+
+## 0.0.11
+
+- feat: support `pathAliases`, "synthetic" routes which gets redirected to canonical path
+
 ## 0.0.10
 
 - Require